+++ Insurance-System/src/main/java/org/example/business/education/BusinessEducationTeam.java
@@ -1,10 +1,10 @@
 package org.example.business.education;
 
 import org.example.Team;
-import org.example.business.design.controller.dto.Request;
-import org.example.business.design.controller.dto.Response;
 import org.example.business.education.model.Education;
 import org.example.business.education.model.EducationListImpl;
+import org.example.common.dto.RequestDto;
+import org.example.common.dto.ResponseDto;
 
 /**
  * @author USER
@@ -21,28 +21,29 @@ public class BusinessEducationTeam extends Team {
 	}
 
 	@Override
-	public void manage() {
-
-	}
-
-	@Override
-	public Response process(int function, Request request) {
+	public ResponseDto manage(RequestDto request) {
 		return null;
 	}
 
 	@Override
-	public void register() {
-
-	}
-
-	@Override
-	public void remove() {
-
-	}
-
-	@Override
-	public Object retrieve() {
+	public ResponseDto process(RequestDto request) {
 		return null;
 	}
+
+	@Override
+	public ResponseDto register(RequestDto request) {
+		return null;
+	}
+
+	@Override
+	public ResponseDto remove(RequestDto request) {
+		return null;
+	}
+
+	@Override
+	public ResponseDto retrieve(RequestDto request) {
+		return null;
+	}
+
 
 }
