+++ Insurance-System/src/main/java/org/example/complaint/model/ComplaintList.java
@@ -1,7 +1,5 @@
 package org.example.complaint.model;
 
-import org.example.complaint.Complaint;
-
 public interface ComplaintList {
     public void add(Complaint complaint);
