+++ Insurance-System/src/main/java/org/example/damageAssessmentTeam/DamageAssessmentTeam.java
@@ -1,9 +1,8 @@
 package org.example.damageAssessmentTeam;
 
 import org.example.Team;
-import org.example.business.design.controller.dto.Request;
-import org.example.business.design.controller.dto.Response;
-import org.example.business.design.view.DesignInspectionController;
+import org.example.common.dto.RequestDto;
+import org.example.common.dto.ResponseDto;
 
 /**
  * @author USER
@@ -15,38 +14,34 @@ public class DamageAssessmentTeam extends Team {
 
 	}
 
+	@Override
+	public ResponseDto manage(RequestDto request) {
+		return null;
+	}
+
+	@Override
+	public ResponseDto process(RequestDto request) {
+		return null;
+	}
+
+	@Override
+	public ResponseDto register(RequestDto request) {
+		return null;
+	}
+
+	@Override
+	public ResponseDto remove(RequestDto request) {
+		return null;
+	}
+
+	@Override
+	public ResponseDto retrieve(RequestDto request) {
+		return null;
+	}
+
 	public void finalize() throws Throwable {
 		super.finalize();
 	}
 
-	public void manage(){
-
-	}
-
-	@Override
-	public Response process(DesignInspectionController.Function function, Request request) {
-		return null;
-	}
-
-//	@Override
-//	public void process(int processNum, int domainId) {
-//
-//	}
-
-	public void process(){
-
-	}
-
-	public void register(){
-
-	}
-
-	public void remove(){
-
-	}
-
-	public Object retrieve(){
-		return null;
-	}
 
 }
